Added check to appsAway_cleanupCluster.sh. Changed one input option.
In this commit the check_params function has been added to notify the user if he puts a wrong input or if he does not put any. Moreover, the string 'all' has been used to handle the total clean up of the machine, instead of not passing any inputs. Lastly, in order to not force the user to use the quotes when passing arguments to the script, $1 has been changed for $@.